Move ContentAccess model into rpc/shared
